Add id to tokens and preload all page defs
For the web this happens as the api returns values and for books and
movies at content load. All the token ids are collected and then all ids
are searched for in the definitions table, and then loaded in the
interface into a window-local variable. This means that it is no longer
necessary to have pinyin in the export/api, but it still doesn't require
a query to the browser db. This also means a browser db query isn't
required for simplers syns, or for getting the token details.

This means that token details will always be instantaneous, even if the
SW has been unloaded.

Signed-off-by: Anton Melser <anton@transcrob.es>