Fix import/content analysis stats
Currently imports can contain "rubbish", in the sense that the parsing
of a text (in Chinese, say) might include pinyin or other English words
(or even non-sense strings) that are *not* classed as "Foreign Words".
This means that they were included in the "words you don't know yet" for
the stats, particularly as they can't be included as definition items!

So at least for the stats, don't include them. Is there any point in
including them in the initial import analysis... Maybe?

Signed-off-by: Anton Melser <anton@transcrob.es>