Activate proper offline support
Unfortunately, the NYPL offline support is fairly completely broken...
It possibly works if you hold it right on their platform, when there is
nothing else being cached or done in the SW. But using the off the shelf
is TOTALLY broken for us. So lots and lots of copy paste... They may try
and improve but we may just need to try and evolve things on our own.

Signed-off-by: Anton Melser <anton@transcrob.es>