Add filestats data for generated .data.js files
This allows normal disk caching. This might be a bad idea, as we have an
explicit caching system. If so, then the non data files will need to
have this *removed* from their headers, so they aren't cached either. As
it stood (before this commit), you could see the .html files but there
was no data file accessible, meaning it looked like it didn't work. So
either have both or neither - currently both get cached (they are very
        unlikely to change for normal users, so why not...)

Signed-off-by: Anton Melser <anton@transcrob.es>