Remove useless copy/pasted code from web-reader
It was unclear (to me anyway) that it was possible to import things that
weren't exported at the root of the web-reader package. This caused a
lot of pseudo bugs to be found...

There may still be a but, as their initSW method doesn't actually ever
attempt to provide the cached publication files - the code has been
commented out. The docs are *absolutely* not clear that this should be
implemented by the user, if that was their intention in leaving this
commented. As it stands, it looks like a bug.

In any case, in order to maintain better control of the absolutely
critical SW infra in this project and as it is unlikely to cause huge
disruption if they change how they do stuff internally, the decision was
made to reproduce the appropriate parts of the logic rather than
directly reuse the code (which didn't work as it stands anyway).

Signed-off-by: Anton Melser <anton@transcrob.es>