Another omnibus commit...
Migrate from faust to a normal agent + separate stats postgres db for
the stats. This was due to the faust streaming project also being dead
and issues with rebuilding the rocksdb. Faust will remain as a task
scheduler until celery properly supports kafka.

Fix and reapply websocket updates for the graphql sync. Migrate from the
abandoned subscriptions-transport-ws to graphql-ws and fix the 10s of
small bugs that stopped it from working properly. Seems to work quite
well now!

A number of small bugfixen and deps updates

Signed-off-by: Anton Melser <anton@transcrob.es>