Better glossing for probably known words
Certain parsing errors/non-dict words are pretty regular, like doubled words (eg. 开开心心), or English punctuation that gets tacked onto the end of words. We could try and reparse but for the moment we'll just do it at gloss presentation time, and look for what should have been parsed, and only show a gloss if *that* word isn't known

+ deps updates
+ a horrible subs refresh bug...

Signed-off-by: Anton Melser <anton@transcrob.es>